upload logic: write to temporary file first
The response used to be written directly into the requested file, now it
is written into a temporary file first which is then moved to the
correct location. This is done to prevent a possible error when two or
more clients try to upload something into the same file at the same time.